v2/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0,<3.0
transformers>=4.48
torch>=2.7
unstructured[all-docs]>=0.12.6
spacy[transformers]>=3.8
networkx>=3.2
rank_bm25>=0.2
numpy>=1.26
nltk>=3.8
matplotlib>=3.8
docx
PyMuPDF==1.25.5